Minor fix - day of the week in the UI discrepancy
I just noticed that the day of the week display on the local UI (LCD)
can be incorrect when the time zone difference leads to the date
difference. Fixed it.

The problem is caused by the difference between the local time (as in
now() function) and the time zone-adjusted time.